Fix exception handler when id not found

+++ src/main/java/com/webservicespring/project/services/UserService.java
@@ -37,10 +37,12 @@ public class UserService {
     }
 
     public void delete(Long id) {
+
+        if (!userRepository.existsById(id)){
+            throw new ResourceNotFoundException(id);
+        }
         try {
             userRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new ResourceNotFoundException(id);
         } catch (DataIntegrityViolationException e) {
             throw new DatabaseException(e.getMessage());
         }
